make changes for FreeBSD 13
Routing rewrite in FreeBSD 13 broke some assumptions

a) route notifications for CARP are different than before; however freebsd 12 and 13 both announce addresses, so this should work both places

b) there's no longer a constraint on a single default route, so change first and add if not found instead of the other way around. This should also work in 12.